requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.9"
dj-config-url==0.1.1
dj-database-url==1.0.0
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pytz==2022.1
Serializer==0.2.1
simplejson==3.17.6
sqlparse==0.4.2
tzdata==2022.1
